Remove unnecessary utils
